backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
boto3==1.39.12
botocore==1.39.12
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.9.2
cron-descriptor==1.4.5
Django==4.2.23
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.4.2
flake8==7.3.0
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
